fix(entity-table): fix errors with existing child entities from one-to-one relationships
Child entities linked to a parent via a one-to-one or one-to-one-or-none relationship must not be present in the database at insertion time, since an entity on the right-hand side of the relationship can only be linked to an entity on the left-hand side of the relationship.

We currently do not respect this consistency on the constraint and columns side of things: in a true one-to-one relationship the RHS entity is bound to the LHS entity via a FK on the RHS side. That way, a RHS entity can refer to one and only one entity on the LHS table. However, in the bind() function for one-to-one relationships, we expect a column on the LHS table to be a FK to the uuid column on the RHS table.

We do not even require the column on the LHS table to be a FK to the RHS table.